Create digits01.cpp
Just some test code, proof of concept for a numeric puzzle. The idea is to find a 10-digit number that contains each of the digits 0-9, such that the number formed by the first digit is divisible by 1 (trivially true), the number formed by the first 2 digits is divisible by 2, and so on for all ten digits. This code just demonstrates a means to test given numbers (here, over a test range that includes a solution). Code to generate the candidate numbers (hopefully in a smarter fashion than just brute force iteration over 10 billion candidates) is left as an exercise.